[WIP&refrac] Re-organize code
