+++ bert_multitask_learning/metrics.py
@@ -4,7 +4,7 @@ from copy import copy
 from nltk.translate import bleu_score
 
 from .input_fn import predict_input_fn
-from .utils import PREDICT
+from .special_tokens import PREDICT
 
 
 def get_ner_fmeasure(golden_lists, predict_lists, label_type="BMES"):
